+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # spring-data-redis-starter
 
-# change log
+## 12-JUNE-2021
+
+### Implementation
+### Implementation
+
+    - Added hateoas backlinks for create todo
 
 ## 09-JUNE-2021